fix(displays): batch configuration apply, transform, mode changes, mirrorOf (#12)
* fix(displays): setMirrorOf logic and pos calc on mirror, add mirroring support in OutputTargetState

ensure mirror cleared on output target state when setting relative to, vise-versa when setting relative to (should clear mirror)

fix(displays): don't error on apply when we have no m_calculation_result, just perform calculate

chore(displays): drop dead code

* fix: batch configuration apply, transform, mode changes

Standardize on mHz (I have gone back and forth on this), as swapping between the two is error prone.

Fixes #8 as well

+++ src/dbus/OutputModeService.cpp
@@ -32,10 +32,10 @@ int OutputModeService::Height() const {
     if (size) return size->height();
     return 0;
 }
-double OutputModeService::RefreshRate() const {
+qulonglong OutputModeService::RefreshRate() const {
     auto refresh = m_mode->getRefresh();
-    if (refresh) return refresh.value();
-    return 0.0;
+    if (refresh) return static_cast<qulonglong>(refresh.value());
+    return 0;
 }
 bool OutputModeService::Preferred() const {
     auto preferred = m_mode->isPreferred();
@@ -61,7 +61,7 @@ QVariantMap OutputModeService::GetModeInfo() {
     auto preferred = m_mode->isPreferred();
     info["Width"] = size ? size->width() : 0;
     info["Height"] = size ? size->height() : 0;
-    info["RefreshRate"] = refresh ? refresh.value() : 0.0;
+    info["RefreshRate"] = refresh ? static_cast<qulonglong>(refresh.value()) : static_cast<qulonglong>(0);
     info["Preferred"] = preferred ? preferred.value() : false;
     info["Current"] = Current();
     return info;
